Skip abstract classes, interfaces and non-public methods
This version is not intended to be a comprehensive API documentation, but rather a usage documentation for developers using the package/library.
Therefore I only want to list public methods in concrete classes and skip all the rest to create a much simplified and more readable documentation than a full blown API doc.